Feature #39 Finish
Tab button color change to be done after
