+++ SalonApp/SalonApp/Controllers/General/artistDetail/Comment.swift
@@ -1,0 +1,14 @@
+//
+//  Comment.swift
+//  SalonApp
+//
+//  Created by engin gülek on 28.08.2023.
+//
+
+import Foundation
+
+struct Comment {
+    let id : Int
+    let nameSurname : String
+    let comment:String
+}
